% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_output}
\alias{get_output}
\title{Import samplesim results}
\usage{
get_output(name, path = ".", change = FALSE, reference = NULL)
}
\arguments{
\item{name}{a character of length 1 giving the name of the simulation.
If \code{NULL} (default) the simulation will be named by the time of the
simulation. This name will serve to create a directory in which results
will be stored.}

\item{path}{a character string of length 1. The directory to save results.
This directory must exist and can be an absolute or a relative path.
Default is the current working directory.}

\item{change}{a logical value. If \code{TRUE} values are expressed as a
percentage of change compared to a reference. Default is \code{FALSE}.}

\item{reference}{an integer. The reference state to compute percentage of
change. If \code{NULL} (default) the minimum value of sample size will be used.}
}
\value{
A data frame with five columns:
\itemize{
\item \code{source}: the source name.
\item \code{size}: the sample size.
\item \code{replicate}: the number of replicate.
\item \code{value}: the value of medians/widths of credible intervals.
\item \code{type}: the label of value (i.e. medians or widths)
}

If \code{change = TRUE}, then the column replicate is omitted and results are
aggregated over replicates.
}
\description{
This function imports samplesim simulation results. More specifically it
loads the median and width of credible intervals.
}
\examples{
\dontrun{
## Please visit the vignette available at: 
## https://ahasverus.github.io/samplesim/articles/samplesim.html
}
}
\seealso{
\code{\link[=samplesim]{samplesim()}}, \code{\link[=plot_samplesim]{plot_samplesim()}}, \code{\link[=plot_isospace]{plot_isospace()}}
}
