% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_samplesim.R
\name{plot_samplesim}
\alias{plot_samplesim}
\title{Plot samplesim simulation results}
\usage{
plot_samplesim(name, path = ".", change = FALSE, reference = NULL)
}
\arguments{
\item{name}{a character of length 1 giving the name of the simulation.
If \code{NULL} (default) the simulation will be named by the time of the
simulation. This name will serve to create a directory in which results
will be stored.}

\item{path}{a character string of length 1. The directory to save results.
This directory must exist and can be an absolute or a relative path.
Default is the current working directory.}

\item{change}{a logical value. If \code{TRUE} values are expressed as a
percentage of change compared to a reference. Default is \code{FALSE}.}

\item{reference}{an integer. The reference state to compute percentage of
change. If \code{NULL} (default) the minimum value of sample size will be used.}
}
\description{
This plot function is a graphical representation of the effects of sample
size on estimates and precision of stable isotope mixing solutions. Two
plots are currently available: a plot for the width of the credible interval
and one for the median of the posterior distribution, both displayed for
each sample size and each source.
}
\details{
This plot function automatically loads results data stored by the
\code{\link[=samplesim]{samplesim()}} function from the simulation name. It produces two plots.
The first plot presents the width of the credible interval displayed for
each sample size and each source. The second represents the median of the
posterior distribution for each sample size and each source.
}
\examples{
\dontrun{
## Please visit the vignette available at: 
## https://ahasverus.github.io/samplesim/articles/samplesim.html
}
}
\seealso{
\code{\link[=samplesim]{samplesim()}}, \code{\link[=get_output]{get_output()}}, \code{\link[=plot_isospace]{plot_isospace()}}
}
