% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesim-package.R
\docType{package}
\name{samplesim-package}
\alias{samplesim}
\alias{samplesim-package}
\title{samplesim: Sample Size Effects in Stable Isotope Mixing Solutions}
\description{
Allows the investigation of the effect of sample size on estimates
    and precision of stable isotope mixing solutions calculated with the 
    package 'siar' (written by Andrew Parnell et al.) or the package 'MixSIAR' 
    (written by Brian Stock et al.). Sample sizes are modified assuming a 
    normal distribution with a user defined mean and standard deviation. 
    Samples of different sizes are created from this distribution, and mixing 
    proportions are estimated for several replicates of each sample size using 
    the function siar::siarmcmcdirchletv4() or the function 
    MixSIAR::run_model(). Authors thank Rolf Ims for his contribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahasverus/samplesim}
  \item Report bugs at \url{https://github.com/ahasverus/samplesim/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicolas Lecomte \email{nicolas.lecomte@umoncton.ca} (\href{https://orcid.org/0000-0002-8473-5375}{ORCID})
  \item Dorothee Ehrich \email{dorothee.ehrich@uit.no}
}

}
\keyword{internal}
