% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_isospace.R
\name{plot_isospace}
\alias{plot_isospace}
\title{Plot data in the isotope space}
\usage{
plot_isospace(
  mix,
  source,
  discr,
  filename = "isospace",
  plot_save_pdf = FALSE,
  plot_save_png = FALSE
)
}
\arguments{
\item{mix}{a named list. The output returned by the
\code{\link[MixSIAR:load_mix_data]{MixSIAR::load_mix_data()}} function and containing consumer isotope values.}

\item{source}{a named list. The output returned by the
\code{\link[MixSIAR:load_source_data]{MixSIAR::load_source_data()}} function and containing mean and standard
deviation isotope values of sources (and in some case raw values).}

\item{discr}{a named list. The output returned by the
\code{\link[MixSIAR:load_discr_data]{MixSIAR::load_discr_data()}} function and containing discrimination values.}

\item{filename}{a character of length 1. The name of the plot file to save.}

\item{plot_save_pdf}{a logical value. If \code{TRUE}, the plot is saved as pdf.}

\item{plot_save_png}{a logical value. If \code{TRUE}, the plot is saved as png.}
}
\description{
This function represents the data in the isotope space to overlay consumer
and sources isotope values. User can also use the \code{\link[MixSIAR:plot_data]{MixSIAR::plot_data()}}.
}
\examples{
\dontrun{
## Please visit the vignette available at: 
## https://ahasverus.github.io/samplesim/articles/samplesim.html
}
}
\seealso{
\code{\link[=samplesim]{samplesim()}}, \code{\link[=get_output]{get_output()}}, \code{\link[=plot_samplesim]{plot_samplesim()}}
}
